separando o projeto em apps
