+++ accounts/urls.py
@@ -4,9 +4,9 @@ from . import views
 
 #teste github actions
 urlpatterns = [
-    path('user/', views.UserRetriveCreate.as_view()),
-    path('user/login/', views.AuthLoginUser.as_view()),
-    path('user/change-password/', views.PasswordUpdate.as_view()),
+    path('', views.UserRetriveCreate.as_view()),
+    path('login/', views.AuthLoginUser.as_view()),
+    path('change-password/', views.PasswordUpdate.as_view()),
+    path('forget-password/', include('django_rest_passwordreset.urls', namespace='password_reset')),
     path('api-auth/', include('rest_framework.urls')),
-    path('user/forget-password/', include('django_rest_passwordreset.urls', namespace='password_reset')),
 ]
