+++ products/admin.py
@@ -1,0 +1,6 @@
+from django.contrib import admin
+from products.models import Brand, Product
+
+# Register your models here.
+admin.site.register(Brand)
+admin.site.register(Product)
