+++ products/serializers.py
@@ -1,0 +1,16 @@
+from rest_framework import serializers
+from products.models import Brand, Product
+
+
+class BrandSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = Brand
+        fields = ['name']
+
+
+class ProductSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = Product
+        fields = ['id', 'name', 'photo', 'description', 'price', 'brand']
+
+    brand = BrandSerializer(read_only=True)
